fix: element regis inside unfocused reference
Before, when you put two nested global components, the overlay is shown
on the wrong element.
This is happening because a nested global element have the same key,
and we're registering all elements.

This commit fix it by disabling element registration, when it's inside
an unfocused element reference. We know whether element reference
is focused or not, by checking whether the elementReferenceDocument
is null.