fix: update localePath to find locales during ISR
There was a weird bug where locale json files were empty for ISR.

I tracked the problem all the way back to the locales being empty in getServerSideProps and when reading through the src for serverSideTranslations there was a lot of `|| {}` code for if the locales can't be found.

Fast forward and while reading through the simple example for the repo I found this in the next-i18next.config.js file

  /** To avoid issues when deploying to some paas (vercel...) */
  localePath:
    typeof window === 'undefined'
      ? require('path').resolve('./public/locales')
      : '/locales',

This ended up solving my problem too.