refactor: split the RuntimeProvider
We had error "Error: This Suspense boundary received an update
before it finished hydrating." when upgrading to Next.js v13.

This was because we were calling setStore before the page finished
hydrating. We fix this by splitting the RuntimeProvider to two
different providers: PreviewProvider, and LiveProvider.

So instead of using an useEffect to rerender the correct provider,
we check for the preview prop passed from the snapshot.