fix: SSR hydration issue with ReactPlayer
See https://github.com/cookpete/react-player/issues/1428 for more
information about why ReactPlayer shouldn't be rendered on the server.
Ideally this is something that would be handled internally by the
ReactPlayer component, but alas, we must take care of it ourselves. We
should replace ReactPlayer with a more robust solution for our built-in
Video component.