fix: update getCurrent before setPropControllers
I ran into an issue where my imperative handle was getting clobbered by
a previous version of getCurrent that had a domNode.

Since this.getCurrent was updated after setPropControllers was called,
setPropControllers was using an older version of "current."

Ultimately, this ensures that editable-text is called with prop
controllers as expected.