fix: remove X and Slack from SocialLinkType
To properly support these types we'll need to version the prop
controller because older versions of the SocialLink component will blow
up if they receive an unknown social link item. Simply adding the type
here would imply that the builder panel can start writing data with X
and Slack social links, but that would be a problem based on what I
described in the sentence before.