feat(next-plugin)!: stop transpiling runtime
We used to need to transpile `@makeswift/runtime` because we were using
`next/dynamic` for code-splitting. We're now using `React.lazy` so it's
not necessary to transpile it anymore. This also makes
`@makeswift/runtime` play better with Next.js' App Router since it's not
treating it as source code and throwing errors for App Router/Pages
Router quirks. For example, importing from `'next/headers'` in Pages
Router.