fix: fix bug in translation of detached typography
Prior to translation we normalize all our `typography` richtext data to the highest ancestor for the simplest possible markup. To do this we go through children determining what parts of `typography` they share. Any `typography` data that is shared is pushed up the tree, while unique `typography` data is left behind.

We do an `and` and an `xor` to determine what is shared and unique. In the case of detached `typography` the logic was incorrect. It was pushing data up the tree wtihout removing it from the original location. This left our normalization in a infinite loop.

My hunch is I added this special case as an optimization, so I am just removing it.