fix: stop destroy socket too early
We were having issues with Arc browser because we were destroying the
HTTP server socket too early. When a request would come in, it would be
immediately canceled. The code snippet seemed to indicate that Chrome
would hold on to the connection, keeping it open, but after testing it,
this didn't seem to be the case. For this reason, I removed the code
snippet since it didn't seem to have any beneficial effects at the
moment.