fix: global element not selectable after it's blurred
This bug only happened on component that doesn't use forwardRef.

This happened because we're relying on FindDomNode when the component
does not use forwardRef. Unfortunately, for ElementReference, we render
the ElementData, instead of Element, which has the FindDomNode fallback.

This commit fix that issue, and we render Element instead of ElementData.