feat: send REGISTER_BUILDER_DOCUMENT during initialize()
This is used to inform the builder which document or element tree to subscribe to.

I considered to reuse the existing REGISTER_DOCUMENT action.
But this action is called within a PreviewProvider useEffect.
I'm afraid that this will caused a race condition issue, if it's
fired before we establish the connection with the builder.