fix: gracefully deserialize controls
Modifies controls deserialization to still continue if an invalid config from
the user is provided. This allows for the panels of successfully created
controls to still appear in the builder. Addresses the case where some users
are not passing valid configs during prop registration, but the issue is not
surfaced due to them most likely not using TypeScript.