fix: stop default rules from overriding props
When using a Box as the root for an embedded component, the width/margin prop
panels do not apply to the Box as expected, even though the data is saved
correctly from the panels. The reason for this is that the CSS class being
generated via the prop data is injected earlier in the stylesheet than the class
for the default styles, resulting in only the latter being applied. This is not
true for when the page root is present around the box, which will result in the
defaults getting injected first, followed by the props class.

Since responsive styles for width/margin are always passed to the
BackgroundsContainer, this fix addresses the problem by only applying the
generated classname for default width/margin if a classname was not passed.