refactor: continuous builder document registration
Currently, the runtime only sends builder documents during the initialization
of the builder <-> host connection. This means that any functionality that could
rely on a remount of a document after the connection has already been
established will not work, since the registration will not be communicated to
the builder.

This change addresses this by adding a document registration hook for embedded
components that sends a builder registration action on mount.