refactor(runtime): pass initial root element directly
Rather than passing metadata about whether the root element exists or not, we
now allow the root element to be null, and also pass an initial root element
field. This allows us to derive whether the root exists, as well as use this
initial root to reset the data later.