fix: fix incorrect types of Font control definition due to TS quirks
When I tried using the Value and Data types of the new Font control within our app they were overly restrictive to the types of the control when `variants: false`. After creating a repro we found that using `| undefined` on methods within our base `Definition` class was causing these types to be incorrect, and that creating a type utility called `UndefinedOr` fixed the issue. We considered calling this `Optional` but that nomenclature is particular to parameter and we wanted something that would also work for return types.

Co-authored-by: Aleksey Gurtovoy <agurtovoy@acm.org>