fix: cache key mismatch when default locale is passed to getPageSnapshot
When we read from the API resource reducer, we're using the locale from `useDocumentLocale`, which is coming from `snapshot.document.locale`. We have a legacy behavior in our API server where if the `locale === defaultLocale`, the server will return the locale as `null`.

When we write to the API resource reducer, we're using the `cacheData` from the introspect function, which gets the locale from `getPageSnapshot` argument directly, without checking if it's the defaultLocale or not.

This commit fix the issue by using the locale returned from the snapshot, instead from the getPageSnapshot argument, to make sure there is no locale mismatch on the document and on the cacheData.