feat: introduce native draft mode
Remove proxying functionality from the runtime in favor of setting draft mode
and version cookies on the client. Draft mode is now set via redirect. In the
case a user is able to set draft cookies on the live site, this feature
introduces a toolbar UI that allows the user to exit the preview.