fix: avoid document context rerenders
As part of the document context, we were passing the document root element. This
means that every edit would result in the context value changing, resulting in
all children rerendering. For larger element trees, this means that all props in
the document would be re-resolved for any prop change. This fix only passes the
necessary fields for the document (key and locale).