fix: authorization error on publish
Fixes an error when the release Github workflow runs. The error is "an error
occurred while publishing makeswift: ENEEDAUTH This command requires you to be
logged in to https://registry.npmjs.org".

According to https://github.com/npm/cli/issues/6184#issuecomment-1544496444,
the fix is to specify a registry url when setting up Node.